разделил код юзера и всего остального
