+++ src/backend/api/urls.py
@@ -1,11 +1,13 @@
 from django.urls import include, re_path
 from rest_framework.routers import DefaultRouter
 
-from api.views import (
-    UserViewset,
+from api.views.towtruck import (
     OrderViewset,
     FeedbackViewset,
+    TariffViewset,
+    CarTypeViewset
 )
+from api.views.users import UserViewset
 
 app_name = "api"
 
@@ -14,6 +16,8 @@ router = DefaultRouter()
 router.register("user", UserViewset, basename="user")
 router.register("order", OrderViewset, basename="order")
 router.register("feedback", FeedbackViewset, basename="feedback")
+router.register("tariff", TariffViewset, basename="tariff")
+router.register("cartype", CarTypeViewset, basename="cartype")
 
 urlpatterns = [
     re_path(r"^", include(router.urls)),
