+++ src/backend/requirements.txt
@@ -1,44 +1,53 @@
 asgiref==3.7.2
+cachetools==5.3.2
 certifi==2023.7.22
 cffi==1.16.0
 cfgv==3.4.0
-chardet==5.2.0
 charset-normalizer==3.3.2
 cryptography==41.0.5
 defusedxml==0.8.0rc2
 distlib==0.3.7
 Django==4.2.7
 django-filter==23.3
+django-gmailapi-backend==0.3.2
+django-phonenumber-field==7.2.0
 django-templated-mail==1.1.1
 djangorestframework==3.14.0
 djangorestframework-simplejwt==5.3.0
 djoser==2.2.1
 drf-yasg==1.21.7
 filelock==3.13.1
-flake8==6.1.0
+google-api-core==2.14.0
+google-api-python-client==2.108.0
+google-auth==2.23.4
+google-auth-httplib2==0.1.1
+googleapis-common-protos==1.61.0
 gunicorn==20.1.0
+httplib2==0.22.0
 identify==2.5.31
 idna==3.4
 inflection==0.5.1
-mccabe==0.7.0
 nodeenv==1.8.0
 oauthlib==3.2.2
 packaging==23.2
+phonenumbers==8.13.25
+Pillow==10.1.0
 platformdirs==3.11.0
 pre-commit==3.5.0
-psycopg==3.1.12
-psycopg2==2.9.9
+protobuf==4.25.1
 psycopg2-binary==2.9.9
-pycodestyle==2.11.1
+pyasn1==0.5.1
+pyasn1-modules==0.3.0
 pycparser==2.21
-pyflakes==3.1.0
 PyJWT==2.8.0
+pyparsing==3.1.1
 python-dotenv==1.0.0
 python3-openid==3.2.0
 pytz==2023.3.post1
 PyYAML==6.0.1
 requests==2.31.0
 requests-oauthlib==1.3.1
+rsa==4.9
 social-auth-app-django==5.4.0
 social-auth-core==4.5.0
 sqlparse==0.4.4
@@ -47,4 +56,3 @@ tzdata==2023.3
 uritemplate==4.1.1
 urllib3==2.0.7
 virtualenv==20.24.6
-gunicorn==20.1.0
