+++ src/infra/nginx.conf
@@ -13,19 +13,19 @@ server {
         proxy_pass http://backend:8000/swagger/;
     }
 
-
     location /api/ {
         proxy_set_header Host $host;
         proxy_pass http://backend:8000/api/;
-    }
-    location /admin/ {
-        proxy_set_header Host $http_host;
-        proxy_pass http://backend:8000/admin/;
     }
 
     location /auth/ {
         proxy_set_header Host $http_host;
         proxy_pass http://backend:8000/auth/;
+    }
+
+    location /admin/ {
+        proxy_set_header Host $http_host;
+        proxy_pass http://backend:8000/admin/;
     }
 
     location / {
